+++ config/config.exs
@@ -23,6 +23,12 @@ config :logger, :console,
   format: "$time $metadata[$level] $message\n",
   metadata: [:request_id]
 
+# Configures Swoosh email adapter
+config :phx_bb, PhxBb.Mailer,
+  adapter: Swoosh.Adapters.Mailjet,
+  api_key: {:system, "MAILJET_API_KEY"},
+  secret: {:system, "MAILJET_SECRET_KEY"}
+
 # Use Jason for JSON parsing in Phoenix
 config :phoenix, :json_library, Jason