+++ lib/phx_bb/mailer.ex
@@ -1,0 +1,4 @@
+defmodule PhxBb.Mailer do
+  @moduledoc false
+  use Swoosh.Mailer, otp_app: :phx_bb
+end