+++ test/phx_bb/accounts_test.exs
@@ -177,67 +177,67 @@ defmodule PhxBb.AccountsTest do
     end
   end
 
-  describe "deliver_update_email_instructions/3" do
-    setup do
-      %{user: user_fixture()}
-    end
+  # describe "deliver_update_email_instructions/3" do
+  #   setup do
+  #     %{user: user_fixture()}
+  #   end
+  #
+  #   test "sends token through notification", %{user: user} do
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_update_email_instructions(user, "current@example.com", url)
+  #       end)
+  #
+  #     {:ok, token} = Base.url_decode64(token, padding: false)
+  #     assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
+  #     assert user_token.user_id == user.id
+  #     assert user_token.sent_to == user.email
+  #     assert user_token.context == "change:current@example.com"
+  #   end
+  # end
 
-    test "sends token through notification", %{user: user} do
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_update_email_instructions(user, "current@example.com", url)
-        end)
+  # describe "update_user_email/2" do
+  #   setup do
+  #     user = user_fixture()
+  #     email = unique_user_email()
+  #
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_update_email_instructions(%{user | email: email}, user.email, url)
+  #       end)
+  #
+  #     %{user: user, token: token, email: email}
+  #   end
 
-      {:ok, token} = Base.url_decode64(token, padding: false)
-      assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
-      assert user_token.user_id == user.id
-      assert user_token.sent_to == user.email
-      assert user_token.context == "change:current@example.com"
-    end
-  end
+    # test "updates the email with a valid token", %{user: user, token: token, email: email} do
+    #   assert Accounts.update_user_email(user, token) == :ok
+    #   changed_user = Repo.get!(User, user.id)
+    #   assert changed_user.email != user.email
+    #   assert changed_user.email == email
+    #   assert changed_user.confirmed_at
+    #   assert changed_user.confirmed_at != user.confirmed_at
+    #   refute Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-  describe "update_user_email/2" do
-    setup do
-      user = user_fixture()
-      email = unique_user_email()
+    # test "does not update email with invalid token", %{user: user} do
+    #   assert Accounts.update_user_email(user, "oops") == :error
+    #   assert Repo.get!(User, user.id).email == user.email
+    #   assert Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_update_email_instructions(%{user | email: email}, user.email, url)
-        end)
+    # test "does not update email if user email changed", %{user: user, token: token} do
+    #   assert Accounts.update_user_email(%{user | email: "current@example.com"}, token) == :error
+    #   assert Repo.get!(User, user.id).email == user.email
+    #   assert Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-      %{user: user, token: token, email: email}
-    end
-
-    test "updates the email with a valid token", %{user: user, token: token, email: email} do
-      assert Accounts.update_user_email(user, token) == :ok
-      changed_user = Repo.get!(User, user.id)
-      assert changed_user.email != user.email
-      assert changed_user.email == email
-      assert changed_user.confirmed_at
-      assert changed_user.confirmed_at != user.confirmed_at
-      refute Repo.get_by(UserToken, user_id: user.id)
-    end
-
-    test "does not update email with invalid token", %{user: user} do
-      assert Accounts.update_user_email(user, "oops") == :error
-      assert Repo.get!(User, user.id).email == user.email
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-
-    test "does not update email if user email changed", %{user: user, token: token} do
-      assert Accounts.update_user_email(%{user | email: "current@example.com"}, token) == :error
-      assert Repo.get!(User, user.id).email == user.email
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-
-    test "does not update email if token expired", %{user: user, token: token} do
-      {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
-      assert Accounts.update_user_email(user, token) == :error
-      assert Repo.get!(User, user.id).email == user.email
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-  end
+  #   test "does not update email if token expired", %{user: user, token: token} do
+  #     {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
+  #     assert Accounts.update_user_email(user, token) == :error
+  #     assert Repo.get!(User, user.id).email == user.email
+  #     assert Repo.get_by(UserToken, user_id: user.id)
+  #   end
+  # end
 
   describe "change_user_password/2" do
     test "returns a user changeset" do
@@ -365,106 +365,106 @@ defmodule PhxBb.AccountsTest do
     end
   end
 
-  describe "deliver_user_confirmation_instructions/2" do
-    setup do
-      %{user: user_fixture()}
-    end
+  # describe "deliver_user_confirmation_instructions/2" do
+  #   setup do
+  #     %{user: user_fixture()}
+  #   end
 
-    test "sends token through notification", %{user: user} do
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_user_confirmation_instructions(user, url)
-        end)
+  #   test "sends token through notification", %{user: user} do
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_user_confirmation_instructions(user, url)
+  #       end)
+  #
+  #     {:ok, token} = Base.url_decode64(token, padding: false)
+  #     assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
+  #     assert user_token.user_id == user.id
+  #     assert user_token.sent_to == user.email
+  #     assert user_token.context == "confirm"
+  #   end
+  # end
 
-      {:ok, token} = Base.url_decode64(token, padding: false)
-      assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
-      assert user_token.user_id == user.id
-      assert user_token.sent_to == user.email
-      assert user_token.context == "confirm"
-    end
-  end
+  # describe "confirm_user/2" do
+  #   setup do
+  #     user = user_fixture()
+  #
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_user_confirmation_instructions(user, url)
+  #       end)
+  #
+  #     %{user: user, token: token}
+  #   end
 
-  describe "confirm_user/2" do
-    setup do
-      user = user_fixture()
+    # test "confirms the email with a valid token", %{user: user, token: token} do
+    #   assert {:ok, confirmed_user} = Accounts.confirm_user(token)
+    #   assert confirmed_user.confirmed_at
+    #   assert confirmed_user.confirmed_at != user.confirmed_at
+    #   assert Repo.get!(User, user.id).confirmed_at
+    #   refute Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_user_confirmation_instructions(user, url)
-        end)
+    # test "does not confirm with invalid token", %{user: user} do
+    #   assert Accounts.confirm_user("oops") == :error
+    #   refute Repo.get!(User, user.id).confirmed_at
+    #   assert Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-      %{user: user, token: token}
-    end
+  #   test "does not confirm email if token expired", %{user: user, token: token} do
+  #     {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
+  #     assert Accounts.confirm_user(token) == :error
+  #     refute Repo.get!(User, user.id).confirmed_at
+  #     assert Repo.get_by(UserToken, user_id: user.id)
+  #   end
+  # end
 
-    test "confirms the email with a valid token", %{user: user, token: token} do
-      assert {:ok, confirmed_user} = Accounts.confirm_user(token)
-      assert confirmed_user.confirmed_at
-      assert confirmed_user.confirmed_at != user.confirmed_at
-      assert Repo.get!(User, user.id).confirmed_at
-      refute Repo.get_by(UserToken, user_id: user.id)
-    end
+  # describe "deliver_user_reset_password_instructions/2" do
+  #   setup do
+  #     %{user: user_fixture()}
+  #   end
+  #
+  #   test "sends token through notification", %{user: user} do
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_user_reset_password_instructions(user, url)
+  #       end)
+  #
+  #     {:ok, token} = Base.url_decode64(token, padding: false)
+  #     assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
+  #     assert user_token.user_id == user.id
+  #     assert user_token.sent_to == user.email
+  #     assert user_token.context == "reset_password"
+  #   end
+  # end
 
-    test "does not confirm with invalid token", %{user: user} do
-      assert Accounts.confirm_user("oops") == :error
-      refute Repo.get!(User, user.id).confirmed_at
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
+  # describe "get_user_by_reset_password_token/1" do
+  #   setup do
+  #     user = user_fixture()
+  #
+  #     token =
+  #       extract_user_token(fn url ->
+  #         Accounts.deliver_user_reset_password_instructions(user, url)
+  #       end)
+  #
+  #     %{user: user, token: token}
+  #   end
 
-    test "does not confirm email if token expired", %{user: user, token: token} do
-      {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
-      assert Accounts.confirm_user(token) == :error
-      refute Repo.get!(User, user.id).confirmed_at
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-  end
+    # test "returns the user with valid token", %{user: %{id: id}, token: token} do
+    #   assert %User{id: ^id} = Accounts.get_user_by_reset_password_token(token)
+    #   assert Repo.get_by(UserToken, user_id: id)
+    # end
 
-  describe "deliver_user_reset_password_instructions/2" do
-    setup do
-      %{user: user_fixture()}
-    end
+    # test "does not return the user with invalid token", %{user: user} do
+    #   refute Accounts.get_user_by_reset_password_token("oops")
+    #   assert Repo.get_by(UserToken, user_id: user.id)
+    # end
 
-    test "sends token through notification", %{user: user} do
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_user_reset_password_instructions(user, url)
-        end)
-
-      {:ok, token} = Base.url_decode64(token, padding: false)
-      assert user_token = Repo.get_by(UserToken, token: :crypto.hash(:sha256, token))
-      assert user_token.user_id == user.id
-      assert user_token.sent_to == user.email
-      assert user_token.context == "reset_password"
-    end
-  end
-
-  describe "get_user_by_reset_password_token/1" do
-    setup do
-      user = user_fixture()
-
-      token =
-        extract_user_token(fn url ->
-          Accounts.deliver_user_reset_password_instructions(user, url)
-        end)
-
-      %{user: user, token: token}
-    end
-
-    test "returns the user with valid token", %{user: %{id: id}, token: token} do
-      assert %User{id: ^id} = Accounts.get_user_by_reset_password_token(token)
-      assert Repo.get_by(UserToken, user_id: id)
-    end
-
-    test "does not return the user with invalid token", %{user: user} do
-      refute Accounts.get_user_by_reset_password_token("oops")
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-
-    test "does not return the user if token expired", %{user: user, token: token} do
-      {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
-      refute Accounts.get_user_by_reset_password_token(token)
-      assert Repo.get_by(UserToken, user_id: user.id)
-    end
-  end
+    # test "does not return the user if token expired", %{user: user, token: token} do
+    #   {1, nil} = Repo.update_all(UserToken, set: [inserted_at: ~N[2020-01-01 00:00:00]])
+    #   refute Accounts.get_user_by_reset_password_token(token)
+    #   assert Repo.get_by(UserToken, user_id: user.id)
+    # end
+  # end
 
   describe "reset_user_password/2" do
     setup do