+++ test/support/fixtures/accounts_fixtures.ex
@@ -24,9 +24,9 @@ defmodule PhxBb.AccountsFixtures do
     user
   end
 
-  def extract_user_token(fun) do
-    {:ok, captured} = fun.(&"[TOKEN]#{&1}[TOKEN]")
-    [_, token, _] = String.split(captured.body, "[TOKEN]")
-    token
-  end
+  # def extract_user_token(fun) do
+  #   {:ok, captured} = fun.(&"[TOKEN]#{&1}[TOKEN]")
+  #   [_, token, _] = String.split(captured.body, "[TOKEN]")
+  #   token
+  # end
 end