+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- CI now checks coverage % of the test suite with codecov.io
+- Email verification implemented with Swoosh's Mailjet adapter
+
 ## [0.3.1] - 2021-04-10
 
 ### Added
@@ -78,6 +85,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - User cache for reduced DB queries
 - Page title updates to reflect current view
 
+[Unreleased]: ttps://github.com/APB9785/phxBB/compare/0.3.1...HEAD
 [0.3.1]: https://github.com/APB9785/phxBB/compare/0.3...0.3.1
 [0.3]: https://github.com/APB9785/phxBB/compare/0.2...0.3
 [0.2]: https://github.com/APB9785/phxBB/compare/0.1...0.2